+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Add `:clean :sort` settings option to disable sorting during `clean-ns`.
 - Add `:keep` value to `:ns-indent-blocks-indentation` setting to don't change indentation during `clean-ns`.
+- Deprecate `install-latest-clojure-lsp.sh` in place of `install` new script.
 
 ## 2021.07.01-19.49.02